ai-agent/requirements.txt
fastapi>=0.110.1
uvicorn[standard]>=0.29.0
langchain>=0.1.16
langchain-community>=0.0.34
langchain-core>=0.1.45
langchain-google-genai>=1.0.3
langgraph>=0.0.38
google-search-results>=2.4.2
beautifulsoup4>=4.12.3
requests>=2.31.0
pydantic>=2.7.0
python-dotenv>=1.0.1
